requirements.txt
blinker==1.4
certifi==2018.8.24
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Blogging==1.2.2
Flask-Caching==1.7.2
Flask-Dance==1.1.0
Flask-FileUpload==0.5.0
Flask-Login==0.4.1
flask-marshmallow==0.9.0
flask-paginate==0.5.5
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==20.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
lazy==1.3
Markdown==3.1.1
MarkupSafe==1.0
marshmallow==3.2.2
oauthlib==2.1.0
Pillow==2.2.1
postgres==3.0.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python-slugify==4.0.0
requests==2.22.0
requests-oauthlib==1.0.0
shortuuid==0.5.0
six==1.11.0
SQLAlchemy==1.2.12
SQLAlchemy-Utils==0.33.5
text-unidecode==1.3
urllib3==1.23
URLObject==2.4.3
Werkzeug==0.14.1
WTForms==2.2.1
